Feature : 운영 및 개발기 db 에 대한 데이터 추가 메서드 추가 =========================================================== [기능 추가] - 개발기에서 검증할 때 데이터를 일일이 넣기 어려웠으며   기존 메서드가 테스트에서 사용하는 메서드와 차이가 있어 수정 반영이 번거로웠음 - 테스트에 있는 클래스를 끌고 올 수 없기 때문에 개발기에도 추가적으로 클래스를 클론하여 생성함
[추후 개선 사항]
- 추후 운영기에서 배포 이후에는 depreciated 될 수 있도록 반드시 수정 필요
- 지금도 테스트 유틸 클래스를 수정하게 되면 개발기에 있는 해당 클래스도 함께 수정을 해야하는 제약사항이 있음
