+++ src/main/java/co/pickcake/initutils/testconfig/RuntimeTestDataSize.java
@@ -1,0 +1,15 @@
+package co.pickcake.initutils.testconfig;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class RuntimeTestDataSize {
+
+    private int batch;
+    private int size;
+    private Long alpha;
+    private Long beta;
+
+}
